feat: detailed card wip
