+++ .env
@@ -1,0 +1,1 @@
+VITE_API_KEY=241c772e
